+++ scripts/daily-check.sh
@@ -13,12 +13,16 @@ context="/opt/context/texmf-context/"
 git config credential.helper \
     '!f() { echo username=$GITHUB_USER; echo "password=$GITHUB_TOKEN"; };f'
 
+# The suffix that we'll use for updates to the TeX Live package that use the
+# same version of ConTeXt
+suffix="A"
+
 # Get the current version
 version="$( \
     grep -oP '(?<=def\\contextversion\{)(\d{4}\.\d{2}\.\d{2} \d{2}:\d{2})' \
     $context/tex/context/base/mkxl/context.mkxl \
 )"
-version="$(echo "$version" | tr '.: ' '-')"
+version="$(echo "$version" | tr '.: ' '-')-$suffix"
 
 # Tag the current version, and exit if it already exists
 git tag --no-sign "$version" || exit 0